turn sample into a multiplatform project
