--- sample/back/build.gradle
@@ -1,51 +1,0 @@
-buildscript {
-    repositories { jcenter() }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-serialization:$kotlin_version"
-    }
-    ext {
-        ktor_version = '1.1.4'
-        kotlin_serialization_version = '0.11.0'
-    }
-}
-
-apply plugin: 'io.gitlab.arturbosch.detekt'
-apply plugin: 'kotlin'
-apply plugin: 'kotlinx-serialization'
-apply plugin: 'application'
-
-group 'com.example'
-version '0.0.1'
-
-repositories {
-    mavenCentral()
-    jcenter()
-    maven { url "https://kotlin.bintray.com/kotlinx" }
-}
-
-dependencies {
-    implementation rootProject  // Local version, otherwise use: compile "io.gladed:watchable:x.x.x"
-    implementation project(':sample:lib')
-
-    implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime:$kotlin_serialization_version"
-    implementation "io.ktor:ktor-server-netty:$ktor_version"
-    implementation "ch.qos.logback:logback-classic:1.2.3"
-
-    testImplementation 'junit:junit:4.12'
-    testImplementation "io.mockk:mockk:$mockk_version"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$kotlin_coroutines_version"
-    testImplementation "io.ktor:ktor-server-test-host:$ktor_version"
-    detektPlugins "io.gitlab.arturbosch.detekt:detekt-formatting:$detekt_version"
-}
-
-compileKotlin.kotlinOptions {
-    jvmTarget = "1.8"
-    freeCompilerArgs += ["-Xuse-experimental=kotlin.Experimental"]
-}
-
-compileTestKotlin.kotlinOptions {
-    jvmTarget = "1.8"
-    freeCompilerArgs += ["-Xuse-experimental=kotlin.Experimental"]
-}
-
-mainClassName = 'MainKt'
