+++ sample/build.gradle
@@ -1,0 +1,114 @@
+buildscript {
+    ext.kotlin_version = '1.3.31'
+    repositories {
+        jcenter()
+    }
+    dependencies {
+        classpath "org.jetbrains.kotlin:kotlin-serialization:$kotlin_version"
+    }
+}
+
+plugins {
+    id 'kotlin-multiplatform'
+}
+apply plugin: 'kotlinx-serialization'
+
+repositories {
+    jcenter()
+    maven { url "https://dl.bintray.com/kotlin/ktor" }
+    mavenCentral()
+}
+
+def ktor_version = '1.1.3'
+def logback_version = '1.2.3'
+def kotlin_serialization_version = '0.11.0'
+
+kotlin {
+    jvm() {
+        compilations.all {
+            kotlinOptions {
+                freeCompilerArgs += ["-Xuse-experimental=kotlin.Experimental"]
+            }
+        }
+    }
+    js() {
+        compilations.all {
+            kotlinOptions {
+                languageVersion = "1.3"
+                moduleKind = "umd"
+                sourceMap = true
+                metaInfo = true
+                freeCompilerArgs += ["-Xuse-experimental=kotlin.Experimental"]
+            }
+        }
+    }
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation kotlin('stdlib-common')
+                api "org.jetbrains.kotlinx:kotlinx-serialization-runtime:$kotlin_serialization_version"
+            }
+        }
+        commonTest {
+            dependencies {
+                implementation kotlin('test-common')
+                implementation kotlin('test-annotations-common')
+            }
+        }
+        jvmMain {
+            dependencies {
+                implementation rootProject  // Local version, otherwise use: compile "io.gladed:watchable:x.x.x"
+                implementation kotlin('stdlib-jdk8')
+                implementation "io.ktor:ktor-server-netty:$ktor_version"
+                implementation "io.ktor:ktor-html-builder:$ktor_version"
+                implementation "ch.qos.logback:logback-classic:$logback_version"
+            }
+        }
+        jvmTest {
+            dependencies {
+                implementation kotlin('test')
+                implementation kotlin('test-junit')
+            }
+        }
+        jsMain {
+            dependencies {
+                implementation kotlin('stdlib-js')
+                implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime-js:$kotlin_serialization_version"
+            }
+        }
+        jsTest {
+            dependencies {
+                implementation kotlin('test-js')
+            }
+        }
+    }
+}
+
+def webFolder = new File(project.buildDir, "../src/jsMain/web")
+def jsCompilations = kotlin.targets.js.compilations
+
+task populateWebFolder(dependsOn: [jsMainClasses]) {
+    doLast {
+        copy {
+            from jsCompilations.main.output
+            from kotlin.sourceSets.jsMain.resources.srcDirs
+            jsCompilations.test.runtimeDependencyFiles.each {
+                if (it.exists() && !it.isDirectory()) {
+                    from zipTree(it.absolutePath).matching { include '*.js' }
+                }
+            }
+            into webFolder
+        }
+    }
+}
+
+jsJar.dependsOn(populateWebFolder)
+
+task run(type: JavaExec, dependsOn: [jvmMainClasses, jsJar]) {
+    main = "MainKt"
+    classpath { [
+            kotlin.targets.jvm.compilations.main.output.allOutputs.files,
+            configurations.jvmRuntimeClasspath,
+    ] }
+    args = []
+}
