--- sample/front/build.gradle
@@ -1,49 +1,0 @@
-group 'com.example'
-version '0.0.1'
-
-buildscript {
-    ext.kotlin_version = '1.3.31'
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-}
-
-apply plugin: 'org.jetbrains.kotlin.frontend'
-apply plugin: 'kotlin2js'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(':sample:lib')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-js:$kotlin_version"
-}
-
-compileKotlin2Js.kotlinOptions {
-    outputFile = "${projectDir}/web/${project.name}.js"
-    moduleKind = "amd" // client-side
-    sourceMap = true
-}
-
-task assembleWeb(type: Sync) {
-    configurations.compile.each { File file ->
-        from(zipTree(file.absolutePath), {
-            includeEmptyDirs = false
-            include { fileTreeElement ->
-                def path = fileTreeElement.path
-                path.endsWith(".js") && (path.startsWith("META-INF/resources/") ||
-                        !path.startsWith("META-INF/"))
-            }
-        })
-    }
-    from compileKotlin2Js.destinationDir
-    into "${projectDir}/web"
-
-    dependsOn classes
-}
-
-assemble.dependsOn assembleWeb
