--- sample/front/src/main/kotlin/Main.kt
@@ -1,3 +1,0 @@
-fun main(args: Array<String>) {
-    println("Hello world!")
-}
