--- sample/lib/build.gradle
@@ -1,60 +1,0 @@
-buildscript {
-    ext {
-        kotlin_version = '1.3.31'
-        kotlinx_serialization_version = '0.11.0'
-    }
-}
-
-apply plugin: 'kotlin-multiplatform'
-apply plugin: 'kotlinx-serialization'
-
-group 'com.example'
-version '0.0.1'
-
-repositories {
-    mavenCentral()
-    jcenter()
-    maven { url "https://kotlin.bintray.com/kotlinx" }
-}
-
-kotlin {
-    jvm()
-    js()
-    sourceSets {
-        commonMain {
-            dependencies {
-                implementation kotlin('stdlib-common')
-                implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime-common:$kotlinx_serialization_version"
-            }
-        }
-        commonTest {
-            dependencies {
-                implementation kotlin('test-common')
-                implementation kotlin('test-annotations-common')
-            }
-        }
-        jvmMain {
-            dependencies {
-                implementation kotlin('stdlib-jdk8')
-                implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime:$kotlinx_serialization_version"
-            }
-        }
-        jvmTest {
-            dependencies {
-                implementation kotlin('test')
-                implementation kotlin('test-junit')
-            }
-        }
-        jsMain {
-            dependencies {
-                implementation kotlin('stdlib-js')
-                implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime-js:$kotlinx_serialization_version"
-            }
-        }
-        jsTest {
-            dependencies {
-                implementation kotlin('test-js')
-            }
-        }
-    }
-}
