+++ sample/src/commonMain/kotlin/Sample.kt
@@ -1,0 +1,9 @@
+expect class Sample() {
+    fun checkMe(): Int
+}
+
+expect object Platform {
+    val name: String
+}
+
+fun hello(): String = "Hello from ${Platform.name}, checkMe=${Sample().checkMe()}"
