+++ sample/src/commonTest/kotlin/SampleTests.kt
@@ -1,0 +1,9 @@
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTests {
+    @Test
+    fun testMe() {
+        assertTrue(Sample().checkMe() > 0)
+    }
+}
