+++ sample/src/jsMain/kotlin/MainJs.kt
@@ -1,0 +1,16 @@
+import kotlin.browser.*
+
+actual class Sample {
+    actual fun checkMe() = 12
+}
+
+actual object Platform {
+    actual val name: String = "JS"
+}
+
+@Suppress("unused")
+@JsName("helloWorld")
+fun helloWorld(salutation: String) {
+    val message = "${hello()} ($salutation from JS)"
+    document.getElementById("js-response")?.textContent = message
+}
