+++ sample/src/jsTest/kotlin/SampleTestsJS.kt
@@ -1,0 +1,9 @@
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsJS {
+    @Test
+    fun testHello() {
+        assertTrue("JS" in hello())
+    }
+}
