+++ sample/src/jvmMain/kotlin/external/Adapter.kt
