+++ sample/src/jvmMain/kotlin/logic/Logic.kt
