+++ sample/src/jvmMain/kotlin/logic/Operations.kt
