+++ sample/src/jvmMain/kotlin/model/Bird.kt
