+++ sample/src/jvmTest/kotlin/SampleTestsJVM.kt
@@ -1,0 +1,9 @@
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class SampleTestsJVM {
+    @Test
+    fun testHello() {
+        assertTrue("JVM" in hello())
+    }
+}
